refactor(checker): move the awaited-type walk into a promise_checker child module
promise_checker.rs reached 2006 lines after #16374's refactor plus this
branch's 34, tripping the 2000-line architecture cap. The awaited-type walk
moves to a child module rather than a sibling so it keeps reading the parent's
private ThenableAwaitInfo and extract_awaited_type_from_valid_thenable, which
stay private.

architecture-check.sh: LOC violations 1 -> 0.
